Refactoring Readme updates Documentation additions
