+++ dormouse-uri/dormouse-uri.cabal
@@ -4,12 +4,23 @@ cabal-version: 1.18
 --
 -- see: https://github.com/sol/hpack
 --
--- hash: d277a67951bf92bff92a7b59de84358b44b62bac1dd4036813970d6f28536d6b
+-- hash: 918ae59b1637bd5f82053e2bb96e041215177ced9aa8a897321889f28d05a5a5
 
 name:           dormouse-uri
 version:        0.1.0.0
-synopsis:       Simple, Type-Safe, Testable HTTP Client
-description:    Safe URI/URL handling
+synopsis:       Library for type-safe representations of Uri/Urls
+description:    Dormouse-Uri provides type safe handling of `Uri`s and `Url`s.
+                .
+                `Uri` sytax is well defined according to [RFC 3986](https://tools.ietf.org/html/rfc3986), Dormouse-Uri parses and encodes `Uri`s according to the syntax defined in this document.
+                .
+                We define `Url` as an absolute URI associated with web resources, the current version of Dormouse-Uri restricts `Url`s to the `http` and `https` schemes.
+                .
+                Dormouse-Uri has the following features:
+                .
+                  - The `Uri` and `Url` data types use `Data.Text` internally, this allows you to freely include percent-decoded characters which will be properly rendered when the `Url`/`Uri` is encoded.
+                  - Quasiquoters to allow safe construction of `Uri`/`Url`s from string literals.
+                  - `DataKinds` allow `Url`s to be restricted to the `http` or `https` schemes are the type level.
+                  - A UrlBuilder syntax to allow type-safe construction/concatenation of `Url`s from their components, e.g. path and query parameters.
 category:       Web
 homepage:       https://github.com/githubuser/dormouse#readme
 bug-reports:    https://github.com/githubuser/dormouse/issues
@@ -41,7 +52,6 @@ library
       Dormouse.Url
   other-modules:
       Dormouse.Uri.Parser
-      Dormouse.Uri.Query
       Dormouse.Uri.Types
       Dormouse.Url.Class
       Dormouse.Url.Types
@@ -71,7 +81,6 @@ test-suite dormouse-uri-test
       Dormouse.Uri.Exception
       Dormouse.Uri.Parser
       Dormouse.Uri.QQ
-      Dormouse.Uri.Query
       Dormouse.Uri.RFC3986
       Dormouse.Uri.Types
       Dormouse.Url
