+++ dormouse-uri/src/Dormouse/Uri/Types.hs
@@ -1,10 +1,8 @@
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE DuplicateRecordFields #-}
-{-# LANGUAGE OverloadedLists #-}
 {-# LANGUAGE GADTs #-}
 {-# LANGUAGE KindSignatures #-}
 {-# LANGUAGE DeriveLift #-}
-{-# LANGUAGE TemplateHaskell #-}
 
 module Dormouse.Uri.Types
   ( UriReference(..)
@@ -28,6 +26,7 @@ import qualified Data.List as L
 import Data.Text (Text, unpack, pack)
 import Language.Haskell.TH.Syntax (Lift(..))
 
+-- | The Username subcomponent of a URI UserInfo
 newtype Username = Username { unUsername :: Text } deriving (Eq, Lift)
 
 instance Show Username where
@@ -36,6 +35,7 @@ instance Show Username where
 instance IsString Username where
   fromString s = Username $ pack s
 
+-- | The Password subcomponent of a URI UserInfo
 newtype Password = Password { unPassword :: Text } deriving (Eq, Lift)
 
 instance Show Password where
@@ -44,11 +44,13 @@ instance Show Password where
 instance IsString Password where
   fromString s = Password $ pack s
 
+-- | The UserInfo subcomponent of a URI Authority
 data UserInfo = UserInfo 
   { userInfoUsername :: Username
   , userInfoPassword :: Maybe Password
   } deriving (Eq, Show, Lift)
 
+-- | The Host subcomponent of a URI Authority
 newtype Host = Host { unHost :: Text } deriving (Eq, Lift)
 
 instance IsString Host where
@@ -57,12 +59,14 @@ instance IsString Host where
 instance Show Host where
   show host = unpack $ unHost host
 
+-- | The Authority component of a URI
 data Authority = Authority 
   { authorityUserInfo :: Maybe UserInfo
   , authorityHost :: Host
   , authorityPort :: Maybe Int
   } deriving (Eq, Show, Lift)
 
+-- | The Fragment component of a URI
 newtype Fragment = Fragment { unFragment :: Text } deriving (Eq, Lift)
 
 instance IsString Fragment where
@@ -71,13 +75,15 @@ instance IsString Fragment where
 instance Show Fragment where
   show fragment = unpack $ unFragment fragment
 
-data UriReference = Absolute | Relative | Unknown
+data UriReference = Absolute | Relative
 
+-- | The Path component of a URI, including a series of individual Path Segments
 newtype Path (ref :: UriReference) = Path { unPath :: [PathSegment]} deriving (Eq, Lift)
 
 instance Show (Path ref) where
   show path = "[" <> L.intercalate "," (fmap show . unPath $ path) <> "]"
 
+-- | An individial Path segment of a URI
 newtype PathSegment = PathSegment { unPathSegment :: Text } deriving (Eq, Lift)
 
 instance IsString PathSegment where
@@ -86,6 +92,7 @@ instance IsString PathSegment where
 instance Show PathSegment where
   show seg = unpack $ unPathSegment seg
 
+-- | The Query component of a URI
 newtype Query = Query { unQuery :: Text } deriving (Eq, Lift)
 
 instance IsString Query where
@@ -94,11 +101,13 @@ instance IsString Query where
 instance Show Query where
   show query = unpack $ unQuery query
 
+-- | The Scheme component of a URI
 newtype Scheme = Scheme { unScheme :: Text } deriving (Eq, Lift)
 
 instance Show Scheme where
   show scheme = unpack . unScheme $ scheme
 
+-- | The data associated with an Absolute URI
 data AbsUri = AbsUri
   { uriScheme :: Scheme
   , uriAuthority :: Maybe Authority
@@ -107,6 +116,7 @@ data AbsUri = AbsUri
   , uriFragment :: Maybe Fragment
   } deriving (Eq, Show, Lift)
 
+-- | The data associated with a Relative URI
 data RelUri = RelUri
   { uriPath :: Path 'Relative
   , uriQuery :: Maybe Query
